// SMP-UI/src/app/models/fnoBhavcopy.ts
export class FNOBhavcopyModel {
    instrument: string;
    symbol: string;
    expiryDate: string;
    strikePrice: number;
    optionType: string;
    open: number;
    high: number;
    low: number;
    close: number;
    settlePrice: number;
    contracts: number;
    valueInLakh: number;
    openInterest: number;
    changeInOpenInterest: number;
    tradedOn: string;
}